+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 
-## 1.1.1 (unreleased)
+## 1.2 (2023-27-04)
 
 Features:
 
-  - Managing issues list -> 
+  - Managing issues list -> [8caae90](https://github.com/michal-stachura/example-contacts/commit/8caae905e882188534581063db6a6cdd10b5c93c)
+    - Add contact form issues list (pt. 8)
+    - Add issue detail with status update form (pt. 9)
+    - Protect issues list, views & status update only for superusers (pt. 10)
 
-## 1.1.0 (2023-25-04)
+## 1.1 (2023-25-04)
 
 Features: